+++ app/src/main/java/dev/dongeeo/netuserspro/data/remote/AuthRemoteDataSource.kt
@@ -1,0 +1,9 @@
+package dev.dongeeo.netuserspro.data.remote
+
+import dev.dongeeo.netuserspro.domain.model.AuthSession
+
+interface AuthRemoteDataSource {
+    suspend fun login(email: String, password: String): AuthSession
+}
+
+